# Use _RecordFunctionFast for torch.autograd.profiler.record_function