--- a/torch/autograd/profiler.py
+++ b/torch/autograd/profiler.py
@@ -781,9 +781,7 @@ class record_function(_ContextDecorator):
         )
 
     def __enter__(self):
-        self.record = torch._C._profiler._RecordFunctionFast(
-            self.name, self.args
-        )
+        self.record = torch._C._profiler._RecordFunctionFast(self.name)
         return self
 
     def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any):