This changes `record_function` to use _RecordFunctionFast, which is more efficient. This is important for Compiler runtime overhead in smaller graphs. Currently this profiler takes half of the overhead.

To measure impact on runtime overhead, here are results from [benchmarks/dynamo/microbenchmarks/overheads.py](https://github.com/pytorch/pytorch/blob/main/benchmarks/dynamo/microbenchmarks/overheads.py).

Before changes:
```
requires_grad=False
compiled 56.9us (warmup=19.3s)

requires_grad=True
compiled 101.6us (warmup=0.2s)

inference_mode()
compiled 53.4us (warmup=0.1s)
```

After changes:
```
requires_grad=False
compiled 27.2us (warmup=9.9s)

requires_grad=True
compiled 61.6us (warmup=0.1s)

inference_mode()
compiled 25.3us (warmup=0.1s)
```