@azahed98 2025-09-30
Scope of this PR is larger than I initially anticipated, and would likely require discussions with others about `record_function` usage and if `_RecordFunction` should be deprecated following this PR's intended change. Closing for now, to be revisited later.